+++ common/structs.py
@@ -676,6 +676,15 @@ class TimelineEvent:
 
 
 @dataclasses.dataclass(slots=True)
+class Review:
+    user: str
+    score: int
+    message: str
+    likes: int
+    timestamp: int
+
+
+@dataclasses.dataclass(slots=True)
 class Label:
     id: int
     name: str
@@ -714,6 +723,7 @@ class Tab:
     name: str
     icon: str
     color: tuple[float] | None
+    position: int
     instances: typing.ClassVar = []
 
     @classmethod
@@ -736,6 +746,17 @@ class Tab:
     def remove(cls, self):
         while self in cls.instances:
             cls.instances.remove(self)
+        cls.sort_instances()
+
+    @classmethod
+    def update_positions(cls):
+        for self_i, self in enumerate(cls.instances):
+            self.position = self_i
+
+    @classmethod
+    def sort_instances(cls):
+        cls.instances.sort(key=lambda self: self.position)
+        cls.update_positions()
 
     @classmethod
     def base_icon(cls):
@@ -860,10 +881,11 @@ class Settings:
     notifs_show_update_banner   : bool
     play_gifs                   : bool
     play_gifs_unfocused         : bool
+    preload_nearby_images       : bool
+    proxy_type                  : ProxyType
     proxy_host                  : str
     proxy_password              : str
     proxy_port                  : int
-    proxy_type                  : ProxyType
     proxy_username              : str
     quick_filters               : bool
     refresh_archived_games      : bool
@@ -978,6 +1000,8 @@ class Game:
     image_url          : str
     previews_urls      : list[str]
     downloads          : tuple[tuple[str, list[tuple[str, str]]]]
+    reviews_total      : int
+    reviews            : list[Review]
     selected           : bool = False
     image              : "imagehelper.ImageHelper" = None
     executables_valids : list[bool] = None
@@ -1153,7 +1177,9 @@ class Game:
             "notes",
             "image_url",
             "previews_urls",
-            "downloads"
+            "downloads",
+            "reviews_total",
+            "reviews",
         ]:
             if isinstance(attr := getattr(self, name), Timestamp):
                 attr.update(value)