+++ main.py
@@ -1,40 +1,13 @@
 #!/usr/bin/env python
+from common import meta
+from modules import patches
+
 import asyncio
 import contextlib
-import os
-import pathlib
 import sys
-
-def mod_version(original_version) -> str:
-    version_parts = original_version.split(".")
-    while len(version_parts) < 3:
-        version_parts.append("0")
-    version_parts[2] = str(int(version_parts[2]) + 1)
-    return ".".join(version_parts)
-
-version = "11.1"
-version = mod_version(version)
-release = False
-build_number = 0
-version_name = f"{version}{'' if release else ' beta'}{'' if release or not build_number else ' ' + str(build_number)} (b-mod)"
-rpc_port = 57095
-rpc_url = f"http://127.0.0.1:{rpc_port}"
-
-frozen = getattr(sys, "frozen", False)
-self_path = pathlib.Path(sys.executable if frozen else __file__).parent
-debug = not (frozen or release)
-
-if not sys.stdout: sys.stdout = open(os.devnull, "w")
-if not sys.stderr: sys.stderr = open(os.devnull, "w")
-if os.devnull in (sys.stdout.name, sys.stderr.name):
-    debug = False
-
-if "F95DEBUG" not in os.environ:
-    os.environ["F95DEBUG"] = str(int(debug))
 
 
 def main():
-    # Must import globals first to fix load paths when frozen
     from modules import globals
 
     from common.structs import Os
@@ -80,7 +53,7 @@ def lock_singleton():
         else:
             try:
                 from urllib import request
-                request.urlopen(request.Request(rpc_url + "/window/show", method="POST"))
+                request.urlopen(request.Request(meta.rpc_url + "/window/show", method="POST"))
             except Exception:
                 pass
 
@@ -93,7 +66,9 @@ def get_subprocess_args(subprocess_type: str):
     return args, kwargs
 
 
-if __name__ == "__main__":
+def _start():
+    patches.apply()
+
     if "-c" in sys.argv:
         # Mimic python's -c flag to evaluate code
         exec(sys.argv[sys.argv.index("-c") + 1])
@@ -119,7 +94,7 @@ if __name__ == "__main__":
                 try:
                     main()
                 except Exception:
-                    if debug and release:
+                    if meta.debug and meta.release:
                         try:
                             from external import error
                             print(error.traceback())
@@ -130,3 +105,7 @@ if __name__ == "__main__":
                         raise
         except KeyboardInterrupt:
             pass
+
+
+if __name__ == "__main__":
+    _start()