+++ modules/db.py
@@ -1,6 +1,7 @@
 import asyncio
 import configparser
 import contextlib
+import dataclasses
 import enum
 import json
 import pathlib
@@ -213,6 +214,7 @@ async def connect():
             "pause_on_429":                f'INTEGER DEFAULT 60',
             "play_gifs":                   f'INTEGER DEFAULT {int(True)}',
             "play_gifs_unfocused":         f'INTEGER DEFAULT {int(False)}',
+            "preload_nearby_images":       f'INTEGER DEFAULT {int(False)}',
             "proxy_type":                  f'INTEGER DEFAULT {ProxyType.Disabled}',
             "proxy_host":                  f'TEXT    DEFAULT ""',
             "proxy_port":                  f'INTEGER DEFAULT 8080',
@@ -308,6 +310,8 @@ async def connect():
             "image_url":                   f'TEXT    DEFAULT ""',
             "previews_urls":               f'TEXT    DEFAULT "[]"',
             "downloads":                   f'TEXT    DEFAULT "[]"',
+            "reviews_total":               f'INTEGER DEFAULT 0',
+            "reviews":                     f'TEXT    DEFAULT "[]"',
         },
         renames=[
             ("executable",           "executables"),
@@ -340,6 +344,7 @@ async def connect():
             "name":                        f'TEXT    DEFAULT ""',
             "icon":                        f'TEXT    DEFAULT "{Tab.base_icon()}"',
             "color":                       f'TEXT    DEFAULT NULL',
+            "position":                    f'INTEGER DEFAULT 0',
         }
     )
     await create_table(
@@ -393,7 +398,10 @@ def sql_to_py(value: str | int | float, data_type: typing.Type):
                     value = data_type([value]) if value else data_type()
                 if args:
                     content_type = args[0]
-                    value = data_type(x for x in (content_type(x) for x in value) if x is not None)
+                    if hasattr(content_type, "__dataclass_fields__"):
+                        value = data_type(x for x in (content_type(**x) for x in value) if x is not None)
+                    else:
+                        value = data_type(x for x in (content_type(x) for x in value) if x is not None)
         case _:
             if isinstance(data_type, types.UnionType):
                 if (
@@ -446,6 +454,7 @@ async def load():
     """)
     for tab in await cursor.fetchall():
         Tab.add(row_to_cls(tab, Tab))
+    Tab.sort_instances()
 
     # Settings need Tabs to be loaded
     cursor = await connection.execute("""
@@ -497,12 +506,16 @@ def py_to_sql(value: enum.Enum | Timestamp | bool | list | tuple | typing.Any):
     elif isinstance(value, list):
         value = value.copy()
         value = [getattr(item, "value", getattr(item, "id", item)) for item in value]
+        if value and hasattr(value[0], "__dataclass_fields__"):
+            value = [dataclasses.asdict(item) for item in value]
         value = json.dumps(value)
     elif isinstance(value, tuple):
         if 3 <= len(value) <= 4 and all(type(item) in (float, int) for item in value):
             value = colors.rgba_0_1_to_hex(value)
         else:
             value = [getattr(item, "value", getattr(item, "id", item)) for item in value]
+            if value and hasattr(value[0], "__dataclass_fields__"):
+                value = [dataclasses.asdict(item) for item in value]
             value = json.dumps(value)
     return value
 
@@ -680,6 +693,7 @@ async def create_tab():
     """)
     tab = row_to_cls(await cursor.fetchone(), Tab)
     Tab.add(tab)
+    Tab.update_positions()
     return tab