+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ### Added:
+- Reviews Tab in More Info Popup (#224 by @Willy-JL)
+  - Labels moved next to executables
 - Image Texture Compression (`(ASTC or BC7) + ZSTD`) option (#212 by @Willy-JL):
   - Compresses images for instantaneous load times (after first compression which is slower)
   - Less VRAM usage, and potentially less disk usage, depending on configuration and GPU support
@@ -6,9 +8,13 @@
 - Unload Images Off-screen option (#212 by @Willy-JL):
   - Saves a lot of VRAM usage by unloading images not currently shown
   - Works best together with Tex Compress, so image load times are less noticeable
+- Preload Nearby Images option (by @Willy-JL):
+  - Starts loading images that aren't visible yet but are less than a window width/height scroll away
+  - Works best together with Tex Compress, so image load times are completely unnoticeable
 - Play GIFs and Play GIFs Unfocused options (#212 by @Willy-JL):
   - Saves a lot of VRAM if completely disabled, no GIFs play and only first frame is loaded
   - Saves CPU/GPU usage by redrawing less if disabled when unfocused, but still uses same VRAM
+- Tabs can not be reordered by dragging (by @Willy-JL)
 
 ### Updated:
 - New notification system with buttons and better platform support, option to include banner image in update notifs (#220 by @Willy-JL)
@@ -20,8 +26,9 @@
 - Fix flashbang while interface is loading (#221 by @sodamouse)
 - Fix GUI redraws not pausing when unfocused, hovered and not moving mouse (by @Willy-JL)
 - Fix missing `libbz2.so` on linux binary bundles (#222 by @Willy-JL)
-- Apply images more efficiently, reduce stutters while scrolling (#212 by @Willy-JL)
+- Apply images more efficiently, eliminate stutters while scrolling, start showing GIFs before all frames are loaded (by @Willy-JL)
 - Improve images error handling and display (#212 by @Willy-JL)
+- Tags now sort alphabetically as expected (by @Willy-JL)
 
 ### Removed:
 - Excluded `libEGL.so` on linux binary bundles, fixes "Cannot find EGLConfig, returning null config" (by @Willy-JL)